backend/requirements.txt
blacksheep
uvicorn
google-genai
google-cloud-storage
google-cloud-tasks
python-dotenv
pydantic-settings
httpx
yt-dlp
asyncpg
openai
cryptography
Pillow
aiohttp
